Bugfixes and Primary loops added
bucketClient merged into main threadingClasses library.
Transcribe function now works async, and no more memory leaks or process overloading!
Added some TODO's.
A lot more work needs to be done, but it currently fires off a job, waits for the job to be complete, and downloads and converts a srt file to a currently hardcoded location. This will be updated in the next push.